+++ src/extract/mod.rs
@@ -41,6 +41,7 @@ pub mod lzop;
 pub mod lzx;
 pub mod mo;
 pub mod mole;
+pub mod mscf;
 pub mod nbh;
 pub mod pdf;
 pub mod pex;